requirements.txt
torch==1.3.1
pyyaml==5.3.1
yamlordereddictloader==0.4.0
tqdm==4.47.0
numpy==1.19.4
scikit-learn==0.22.1
matplotlib==3.1.2
requests==2.24.0
dgl==0.4.3
dglke==0.1.2
pandas
seaborn